[shivani] : Added functionality of finding maximum in given float

+++ src/com/bridgelabz/FindMaximum.java
@@ -2,11 +2,11 @@ package com.bridgelabz;
 
 public class FindMaximum {
 
-	// UC1 -->> Find maximum in given three integer
+	// UC2 -->> Find maximum in given three Float
 
-	public Integer checkMaximum(Integer X, Integer Y, Integer Z) {
+	public Float checkMaximum(Float X, Float Y, Float Z) {
 
-		Integer max = X;
+		Float max = X;
 		if (Y.compareTo(max) > 0) {
 			max = Y;
 		}
@@ -14,7 +14,7 @@ public class FindMaximum {
 			max = Z;
 		}
 		System.out.println(max);
-		return max; // returns the largest number
+		return max;
 	}
 
 }
